add WeatherApiService
